Cleanup pybind declarations, reordering and moving as needed. Add tests for pybind interfaces using pybind11_stubgen. Remove ignore from docs builds.
