+++ python/tests/unit/arcticdb/version_store/test_normalization.py
@@ -18,7 +18,7 @@ import dateutil as du
 import pytest
 import pytz
 from numpy.testing import assert_equal, assert_array_equal
-from arcticdb_ext.version_store import SortedValue as _SortedValue
+from arcticdb_ext.storage import SortedValue as _SortedValue
 
 from arcticdb.exceptions import ArcticDbNotYetImplemented, ArcticException
 from arcticdb.version_store._custom_normalizers import (
