+++ setup.cfg
@@ -130,6 +130,7 @@ Testing =
     psutil
     memray; python_version >= '3.7' and (platform_system == 'Linux' or platform_system == 'Darwin')
     pytest-memray; python_version >= '3.8'and (platform_system == 'Linux' or platform_system == 'Darwin')
+    pybind11-stubgen
 
 [options.entry_points]
 console_scripts =
